ml-model/requirements.txt
# ML Model Training Requirements
tensorflow>=2.13.0
numpy>=1.24.0
pillow>=10.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
tensorflowjs>=4.12.0

